make `idn_connect` available on the lib
