+++ cli/src/main.rs
@@ -15,18 +15,13 @@
  */
 
 #![allow(missing_docs)]
-use subxt::{backend::rpc::RpcClient, client::OnlineClient, config::SubstrateConfig};
-use subxt_signer::sr25519::dev;
-use std::fs::File;
-use std::time::Instant;
 use clap::{Parser, Subcommand};
-use thiserror::Error;
 use etf::runtime_types::node_template_runtime::RuntimeCall::Balances;
+use murmur_lib::{create, etf, idn_connect, prepare_execute, BlockNumber, MurmurStore};
 use sp_core::crypto::Ss58Codec;
-use murmur_lib::{
-    etf, create, prepare_execute, 
-    MurmurStore, BlockNumber
-};
+use std::{fs::File, time::Instant};
+use subxt_signer::sr25519::dev;
+use thiserror::Error;
 
 /// Command line
 #[derive(Parser)]
@@ -52,7 +47,7 @@ struct WalletCreationDetails {
     #[arg(long, short)]
     seed: String,
     #[clap(long, short)]
-    validity: u32
+    validity: u32,
 }
 
 #[derive(Parser)]
@@ -64,7 +59,7 @@ struct WalletExecuteDetails {
     #[arg(long, short)]
     to: String,
     #[arg(short, long)]
-    amount: String
+    amount: String,
 }
 
 #[derive(Error, Debug)]
@@ -114,11 +109,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             println!("✅ MMR proxy account creation successful!");
         }
         Commands::Execute(args) => {
-            let from_ss58 = sp_core::crypto::AccountId32::from_ss58check(&args.to)
-                .unwrap();
+            let from_ss58 = sp_core::crypto::AccountId32::from_ss58check(&args.to).unwrap();
 
             let bytes: &[u8] = from_ss58.as_ref();
-            let from_ss58_sized: [u8;32] = bytes.try_into().unwrap();
+            let from_ss58_sized: [u8; 32] = bytes.try_into().unwrap();
             let to = subxt::utils::AccountId32::from(from_ss58_sized);
             let v: u128 = args
                 .amount
@@ -142,41 +136,14 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             )
             .await;
             // submit the tx using alice to sign it
-            let _result = client.tx()
+            let _result = client
+                .tx()
                 .sign_and_submit_then_watch_default(&tx, &dev::alice())
                 .await;
-        },
+        }
     }
     println!("Elapsed time: {:.2?}", before.elapsed());
     Ok(())
-}
-
-/// async connection to the Ideal Network
-/// if successful then fetch data
-/// else error if unreachable
-async fn idn_connect() -> 
-    Result<(OnlineClient<SubstrateConfig>, BlockNumber, Vec<u8>), Box<dyn std::error::Error>> {
-    println!("🎲 Connecting to Ideal network (local node)");
-    let rpc_client = RpcClient::from_url("ws://localhost:9944").await?;
-    let client = OnlineClient::<SubstrateConfig>::from_rpc_client(rpc_client.clone()).await?;
-    println!("🔗 RPC Client: connection established");
-
-    // fetch the round public key from etf runtime storage
-    let round_key_query = subxt::dynamic::storage("Etf", "RoundPublic", ());
-    let result = client
-        .storage()
-        .at_latest()
-        .await?
-        .fetch(&round_key_query)
-        .await?;
-    let round_pubkey_bytes = result.unwrap().as_type::<Vec<u8>>()?;
-
-    println!("🔑 Successfully retrieved the round public key.");
-
-    let current_block = client.blocks().at_latest().await?;
-    let current_block_number: BlockNumber = current_block.header().number;
-    println!("🧊 Current block number: #{:?}", current_block_number);
-    Ok((client, current_block_number, round_pubkey_bytes))
 }
 
 /// read an MMR from a file
